task_creation_project/task_creation/requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
ipdb==0.13.5
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
# mysqlclient==2.0.3
psycopg2==2.8.6
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
python-http-client==3.3.2
pytz==2021.1
sendgrid==6.6.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
traitlets==4.3.3
wcwidth==0.2.5
openpyxl==3.0.7
requests==2.23.0
boto3
